Converts bulk docs controller to service
